Fixes #1459 - Use SPDX license identifier in package.json files
We are using "BSD" as the license name in our Clay package.json files,
which causes Yarn to emit warnings like this:

    warning package.json: License should be a valid SPDX license expression

We should switch it to "BSD-2-Clause", as listed here on the SPDX
website:

    https://spdx.org/licenses/BSD-2-Clause.html

which corresponds (somewhat loosely) to our LICENSE.md.

Note that because this commit updates:

    packages/generator-metal-clay/app/templates/_LICENSE.md

future generated packages will automatically benefit from this change.

Related: https://issues.liferay.com/browse/IFI-412