Fixes #1457 - Refer to Yarn consistently in Clay README
The Clay README instructs people to install and use Yarn, and we ship
with a yarn.lock, but we don't consistently use `yarn` in the commands
that we show in the README; this commit fixes that for consistency.

Tested the commands manually to make sure they all run as intended.

Related: https://issues.liferay.com/browse/IFI-421