fix(useFocusManagement): remove reset from last position when any list item is deleted
Resetting each time a list item is deleted can cause looping consequences, when you have a component or element that updates the state of something whenever focus or blur occurs, React will throw a ref with null to delete it. element and then launch a new one with it updated. This can be an internal problem with our components if it is non-standard React behavior.

I'm removing so that `moveFocusInScope` corrects the position when it is no longer in the list.