fix(useFocusManagement): sets the last position when one does not exist, with the active element index if it exists in the list
Set the `lastPositionRef` to the active element index is important so that it does not break the focus order. A user may be interacting with an element that in focusManager's position is index 3 but no interaction with focusManager has yet been made, probably when it happens it will start from the first index or second.