chore: re-enable disabled tests
As I noted in this comment:

https://github.com/liferay/clay/pull/2293#issuecomment-523388558

- The snapshots need an update but I don't have enough context to judge
  if the changes are correct.
- The tests trigger an `Error: Not implemented: window.resizeTo`
  message, but don't fail. The original PR (#1884) stubbed `resizeTo` in
  the scripts/setupTests.js, but that didn't make it into the final PR
  (#1978).

So, I am renabling the test, updating the snapshot, and removing the
`window.resizeTo()` calls which currently don't affect the pass/fail
status, and I'll create a ticket for @diegonvs to review the tests and
decide whether any follow-up changes are required. I figure that
passing-but-possibly-not-ideal tests are a small step up from
totally-disabled tests.