fix(@clayui/modal): wrap sub-portals and modal content over a div
Creates a div to aggregate Modal content and all portals attached to Modal, which prevents Modal from closing when there is any interaction within the portals.

Before this fix Modal checks if the elements that will have clicks are inside `modal-dialog`, if true, it does not close the modal and if it is outside, closes the modal.