refactor(@clayui/shared): rewrite useFocusManagement hook to remove use of createScope
The `useFocusManagement` hook rewrite allows to remove the manual tagging API from items that should be added to the scope.

Instead of consuming the DOM directly, now it traverse the Fiber for get focal elements, the Fiber is safer than the DOM since the use of `React.Portal` influences the tree structure and the fiber maintains the tree.