fix: unbreak d3 types
So the linter was complaining about duplicate imports from the "d3"
module, because it can't know that "* as d3" is importing all the
exports, and the other one is importing only types.

I had changed the latter to use destructuring, not realizing that they
were just types, and of course that won't work. Need to put the
duplicate `import` back, and just suppress the lint, because there is no
way to write a single import that imports all the exports *and* also
some types.