fix(@clayui/data-provider): replace setInterval in favor of setTimeout and polling engine improvements
Replace the use of `setInterval` to `setTimeout` is safer for now, because we may have cases when `pollInterval` is set to 1000ms and the request may take longer than this time and we may have more than one request being made. and that number can accumulate. So we will always wait for the request to succeed in scheduling the next one.

Such a change means that `pollInterval` is not always executed every 1000ms for example, it also contains the time of the previous request. In a way, we already had this when we were using `setInterval`... but we are preventing more than one request from being made at the same time and possibly creating some kind of server overload.

A change to `pollInterval` also to enable you to make a mechanism to enable and disable `poll` by setting to some value above 0, enable and equal to 0/false disable.