fix: make sure dependency checker actually walks tree
There are a few differences between the code in:

    https://github.com/liferay/clay/pull/2530

and the original over here:

    https://github.com/wincent/js/blob/master/packages/workspace-scripts/src/dependencies.ts

so I am not sure where things diverged, but the code that walks the
filesystem wasn't actually waiting for callbacks to finish.

Test plan:

1.  Introduce some errors:
    https://gist.github.com/wincent/0e50866606cb38fdaf54680e0e53f4da
2.  `yarn checkDeps` and see errors get reported:
    https://gist.github.com/wincent/82bc4c83e74b47889b4709bb4a60564e
3.  Fix errors and see that some (legit) errors still remain:
    https://gist.github.com/wincent/b09e88aa6dd813bd2a6e7958a4e2fe5d

To keep things clean, I will fix the legit errors in a commit on top of
this one.