test: run build in CI before running checkDeps
The missing dependencies check runs against the built files in "lib/",
so we have to do the build in order to actually run the check. Hopefully
this won't make the CI runs so long that they time out (I don't think it
will because Travis will only kill long-running jobs if they don't
produce any console output for 10 minutes, IIRC, and the build is pretty
chatty).