refactor: fold `yarn build` into `yarn checkDeps`
As discussed here:

https://github.com/liferay/clay/pull/2530#issuecomment-564693121

The `checkDeps` script runs, at least in part, on the built files, which
means that if you don't run a build first you run the risk of having
nothing to actually check, or perhaps checking something stale from a
previous build. This is a bit slow, but if the check is to be robust it
it better to be slow-and-correct than fast-but-possibly-inaccurate.