fix(card): prevent unwanted DOM thrash in ClayCard (#2916)
Not sure if this is a React bug or we were just doing it wrong, but
using the `<Content>` component inside the `ClayCard` implementation was
causing React to remove one of the descendent elements from the DOM and
re-insert it, even though there was no actual diff. When the descendent
element is an `<img>`, this is bad and can cause a visual flicker (maybe
something we haven't noticed before because we tend not to re-render
cards much; but @boton noticed it now because he is rendering a progress
bar inside the card that updates frequently).

To see this in action, see this demo:

https://codesandbox.io/s/blissful-shaw-xrekn

There is a fork of `ClayCard` in there called `ClayCardFork` so that you
can edit it. When you turn off the browser cache, and use the
`<Content>` component, you'll see a new image request going out over the
network every second, even though nothing in the DOM is actually
changing.

If we avoid the `<Content>` component and just inline the
`Context.Provider` directly, no DOM mutations occur and no bogus network
requests happen. Fortunately, the solution actually looks a bit less
repetitive than what we had before, so there is no downside to
this workaround.

Closes: https://github.com/liferay/clay/issues/2916