chore: update for compliance with current Outbound Licensing Policy
This change is analogous to the one I made in:

- https://github.com/liferay/eslint-config-liferay/pull/151
- https://github.com/liferay/liferay-npm-tools/pull/394
- https://github.com/liferay/liferay-js-themes-toolkit/pull/441

It updates the license header format to match the latest policy:

https://grow.liferay.com/excellence/Liferay+Outbound+Licensing+Policy

Did this by updating `copyright.js` and running `yarn lint:fix`. I used
a Vim macro to strip out the original headers. Note some files required
manual fixing because they had ESLint or other comments at the top of
the file (this is bad; the header should always be at the top). Some of
the comments I could just remove (eg. unnecessary lint suppressions),
but the others I moved below the header.

Adds a copy of the license to `LICENSES/MIT.txt` as required by the
policy too.