chore: update lints
Updates to eslint-config-liferay v19.0.2 which brings in some new rules:

- curly (enforces use of curlies around conditional blocks).
- liferay/array-is-array (enforces use of `Array.isArray` instead of
  `instanceof Array`).
- liferay/import-extensions (prohibits superflous extensions in imports).
- liferay/sort-import-destructures (yet another in the family of
  "sorting" rules).
- padding-line-between-statements (specifically, requiring a blank line
  before `return` statements, at Brian's request).
- react/forbid-foreign-prop-types (so that we can be sure that stripping
  propTypes in production builds is safe; obviously won't be relevant in
  this repo).

and produces these new lint errors (about 67 of them):

https://gist.github.com/wincent/84bbd941eec684218acb21ad34fdb156

All of those have autofixes, so I applied them.

I also removed some cruft from the config file (values that are not
needed because they are the defaults).

Note that in testing this, I discovered that the
new "liferay/sort-import-destructures" rule and the
`@typescript-eslint/parser` parser were not playing nicely together. I
first tried updating some of the related NPM packages (ESLint, the
parser etc) but that didn't work (although it did require me to
update one config value), so I then dug deeper and tweaked the rule
implementation to work with both the `@typescript-eslint/parser` and
"espree" (ie. default) parsers:

https://github.com/liferay/eslint-config-liferay/pull/153