fix(@clayui/date-picker): fixes normalization error for the day selected and button to change to the current day
There was a certain problem when using `normalizeDate` it is only used for the month status that we changed the time and the day for the 1st of the month to avoid timezone problems but this was used to normalize the `daysSelected` which caused a problem.

Then it was removed to avoid confusion and we directly use the `setDate` function to normalize only the time from `daysSelected` and the `NEW_DATE` which is used as the default value and to change the selected day to the current date, so we are able to prevent problems where we use the date comparison to check intervals.