refactor(@clayui/shared): don't coerce `contentEditable`
Not sure whether this would actually cause a bug, so labeling this as a
refactor commit.

`contentEditable` can be a string (eg. `'true'`)or a bool (or
`undefined`), so we probably shouldn't coerce it (eg. `!!'false'` is
`true`). Instead, explicitly check against `true` and `'true'`.

Originally started looking at this because the `!!` in this PR looked a
bit off:

    https://github.com/liferay/clay/pull/3942