refactor(@clayui/shared): sort expressions
Seeing as The Liferay Way™ is to sort sortable things, I'm sorting the
expressions in these statements. Sometimes there are logical reasons for
not sorting in order (eg. wanting to short-circuit) or performance
reasons (eg. putting the cheapest checks first), but in this case, the
comparisons all just string equality checks which can be considered to
have essentially zero cost.

Noticed while reading:

    https://github.com/liferay/clay/pull/3942

There is one other sortable thing in that PR that I didn't touch, and
that is `FOCUSABLE_ELEMENTS`, because there is just a tiny little chance
that it might have some subtle performance ramification depending on how
the browser matches the selector, so I'd rather not risk changing it.