feat(@clayui/core): adds multiple selection implementation for Tree View
This implementation is a lightweight and efficient solution for multiple Node selection, it also supports intermediate state and tree selection behaviors if the Node has children.

We maintain a light, flat tree structure and build it along with React's render stream to avoid traversing the tree in a separate stream, so we can have big performance gains on very large trees.