chore(@clayui/core): uses RFC 6902 (JSON Patch) implementation to move item in drop action
This is just a draft for an attempt to implement RFC 6902 (JSON Patch), it also won't cover all the cases that the RFC covers but uses the standards to implement the `move` operation which is what interests us most for perform the `drop` operation, also somewhat based on the Immer API which is used for immutability.

It still has to add the implementation of applying patches and dealing with immutability which is being very difficult to do.

In section 4.4 the `path` and `from` are a string but we are using an array of indices to simplify the case here.

I'm not considering adding the index to tree structure to `items` because that would be an operation we need to call again when implementing "load more" and it's a little expensive depending on the depth and size, just testing other alternatives.