chore(@clayui/core): avoid making calls to setState when moving the item
`open` is called several times when the item is moving, it causes a lot of slowness when moving the item, and frame drop because although we don't add anything to the `expandedKeys` we create a copy of it and we will save it, it schedules a new one task in the React Scheduler, as we do this over and over again, we are flooding with tasks.

fix(@clayui/core): fixes error when getting items from props instead of state

fix(@clayui/core): fix error when starting internal state

As we fixed the behavior of `useInternalState` to follow the behavior of if it doesn't have any of the properties it will use the internal state, we also need to support the initial state, if the initial state is not passed but there is a value in `value` prop, we should use it as state initial.