fix(@clayui/tooltip): forces tooltip to be repositioned when the x-axis is modified
The idea here is to force `align` to recalculate the Tooltip position in a position contrary to the center alignment which can happen the x-axis is modified because the Tooltip is on a smaller screen, this avoids having the x-axis modified by choosing the best position.

This fixes when you are in the top or bottom position.