fix(@clayui/core): fix error when moving item up, down and add to children
chore(@clayui/core): adds more test cases by moving nested items

This covers most use cases when moving an item, moving a nested item, persisting the items, moving to the last item in the list.

fix(@clayui/core): fixes error when moving the item inside the item and its children

chore(@clayui/core): fixes the test of the operation the move item on the object

The behavior is now different, if the indexes are the same size, it means you must move the item to the same level of the hierarchy but if you want to move the item inside an item add a new index 0 to path.

chore(@clayui/core): add the test case by moving the item above and below an item

This case is very simple and doesn't cover all the use cases of moving the item up and down over an item, add in the next interactions:

- Move item above or below a nested item
- Move the item below the last item in the hierarchy
- Move the item above the first item in the hierarchy

fix(@clayui/core): fixes error when adding an empty item in the structure when navigating indexes

This fixes two main issues when using indexes to move items:

- Array can change the position of items when deleting first the item
- Move an item below the last item

chore(@clayui/core): open the item only when hovering over the element for a few ms

When moving an item that has a lot of nesting, opening as it moves the item to the desired place causes a bad experience, especially if the items have many items, use the strategy of opening only if it stays still on the item for a few ms that shows the intention to open the item instead of opening just swiping when you don't have an intention to open.

chore: regen yarn.lock

Some things got lost during the PR rebase with the master.