chore(@clayui/core): does not render the group if it is not expanded
Well, when we have too many items to be rendered this will cause a drop in performance causing the tab to crash, this is because the browser and React itself cannot manage more than 10,000 items and in our tests, we will render more than 100 thousand items.

The problem with this is that it harms the implementation of multi-selection that depends on the items being rendered to mount the elements' layout but we can look at other strategies to resolve this.