chore(@clayui/core): add new `add` operation a value to tree
RFC 6902 implements many operations for a patch a JSON, but we don't need to apply all of them, we'll add them little by little, in this case, add the `add` (4.1) operation to make the `insert` method available, this makes the possibility easier. of creating the implementation of "load more", the return value we can add to the structure easily.