fix(@clayui/css): Atlas and Cadmin Stickers swap background and foreground colors for `sticker-light` and `sticker-dark`
BREAKING CHANGE: `sticker-light` now has a light font color and `sticker-dark` has a dark font color. All instances of `sticker-light` should be updated to use `sticker-dark` and all instances of `sticker-dark` should be updated to use `sticker-light`. If you do not want to make this change in your markup, you can revert back to the original styles with:

```
$sticker-light-bg: $dark !default;
$sticker-light-color: $white !default;

$sticker-dark-bg: $dark !default;
$sticker-dark-color: $white !default;
```