feat(@clayui/css): Globals and Cadmin Globals adds `$enable-clay-color-functions-process-fallback` and `$cadmin-enable-clay-color-functions-process-fallback` variables. This forces Clay color functions to return the CSS Custom Property instead of processing the fallback color and returning a hex value.
feat(@clayui/css): Color Functions adds `clay-color-functions-process-fallback`. A function that checks if the user prefers having Sass process CSS Custom Property fallback colors or if they prefer returning the CSS Custom Property as is.

feat(@clayui/css): Color Functions adds `is-css-var`. A function that checks the value of a Sass variable and returns `true` if the value is a CSS Custom Property.