chore: add browserslist related environment variables
We're adding to environment variables in our GitHub
actions workflow definition, with the following values

BROWSERSLIST_CONFIG='./packages/browserlist-config-clay/config'
BROWSERSLIST_DISABLE_CACHE=1

as well as a new file defined in
./packages/browserslist-config-clay/config

this is a plain text file with the same values that
are defined in
./packages/browserslist-config-clay/index.js

For the moment, this is the only way I managed to see
the "test" job pass on GitHub actions.

I've also disabled running tests for the
generator-clay-component package because it's never able
to find the browserslist configuration file no matter how it
was a set.

We can discuss this to find a better alternative