chore: refactors the `buildIconsSvg` method to use a writable stream
It refactors the implementation of creating the icons.svg file using a writable stream instead of keeping each file in memory and all data too, thus keeping fewer data in memory.

It also closes the process with exit(1) to terminate the process saying that the build failed so that the CI can fail.