chore: fix storybook related dependencies
In https://git.io/JitzN we updated css-loader,
in https://git.io/JitgR we updated sass-loader,
in https://git.io/JitVD we updated style-loader,
and in https://git.io/Jitwf we updated fork-ts-checker-webpack-plugin.
The problem is that these versions are throwing errors
with the current version on Gatsby, so we're reverting
everything to make sure storybook keeps on working
as expected.