chore(@clayui/css): adds pipeline to automatically map all icons into a sass variable
This is a simple implementation that uses JSDom to manipulate the icon and remove some attributes that are needed to add icons to a map in a sass variable, the SVG is also minified.

One strategy we might want to try to decrease the flow of opening icons more than once is to use the same flow as the `buildIconsSvg` method to implement this.