fix(@clayui/provider): removes dependency from ClayModal to avoid circular dependency issues
Unfortunately, this is the only option we have for not using `peerDependencies`. We can for now remove the ClayModalProvider from the ClayProvider and when we can stabilize the package dependencies on `@clayui/core` we can add that support again, it's safe to remove the ClayProvider because it's not used yet.