chore(@clayui/core): select Nodes not visible in DOM/LayoutKeys
This strategy is just a complement to the implementation of multi-selection that creates a layout of the keys following the React rendering flow, in other words, at each Node that is being rendered a map is created, this map is reactive to the component's lifecycle if the component is unmounted it removes the Node from the structure and its references.

This works great for multi-selection to not need to traverse the full tree but if the component is not rendered we have no way to select all the children of a node, in this case this commit implements a secondary strategy like a lazy selection, as far as we have it data of the current Node in the recursive selection of children, if the Node still has children but it is not visible, we do not create the Node layout but use the loc or indexes to get to the Node in the structure items if it is a dynamic TreeView and follow selecting the child recursively.

We don't add these non-visible Nodes to the Layout because the Layout has to be reactive, we need to manage the Node's references and remove it when it no longer exists, either because it was removed or moved, for example. Using the component lifecycle we get this reactivity for free, decreasing complexity.