fix(@clayui/core): remove the context property from DndProvider
When using this component in applications that also use `react-dnd`,
this error is thrown

'Cannot have two HTML5 backends at the same time.'

Here we're just removing the context prop to see if that fixes the
issue. Since there's no way to test this in portal without releasing,
we'll just push this change to a branch for other teams to be able to
test it locally.