chore(@clayui/core): adds support to trigger selection manually
Exposes the selection method to manually trigger via render props to allow customizing in the UI the element that the trigger.

Create a small "protection" layer from the inner method to the exposed method, with the `selection.key` instead of having direct access to the inner structure, in this case, we just want to expose the possibility to check if the element is selected if the UI need to demonstrate this visually to the user.