chore(@clayui/core): adds the `onClick` method of the TreeView.ItemStack being called in the Node event source
Basically, we want to mask that `onClick` is called directly in the ItemStack, ie think of it as syntactic sugar, because internally markup is different from how components are used to abstract the markups and simplify usage.

Adding `onClick` to the TreeView.ItemStack allows the method to cause the `event.preventDefault()` event and ignore the behavior of expanding for example or lazy loading.

We managed to create a design consistency, that is, how the component is used and the perception of where the onClick is being added. The common reasoning when looking at components is that the `TreeView.Item` might be the triggering element but when it has the nested group of elements, the `TreeView.ItemStack` would assume that responsibility because we imagine that's where it is.

Consider this example:

```
<TreeView.Item>
  <Icon symbol="folder" />
  Documents
</TreeView.Item>
```

The component design intuitively implies that the `TreeView.Item` is the trigger here.

Another example:

```
<TreeView.Item>
  <TreeView.ItemStack>
    <Icon symbol="folder" />
    Documents
  </TreeView.ItemStack>
  <TreeView.Group>
    {...}
  </TreeView.Group>
</TreeView.Item>
```

Intuitively the trigger to expand Node is now `TreeView.ItemStack`.

As a common pattern with web and component, being able to prevent the default behavior here would be acceptable and consistent. We can thus be consistent in the component design intuition and abstract the "complexity" of the internal markup.