chore(@clayui/core): add ownership change of properties between item and node
Well in commit 8e03f43d8ad030b6da7695d332d581a2650cb62f we did this by just bringing `onClick` from the ItemStack and the Item to toggle where it should be included.

The same explanations added in the referenced commit description are also applied here. In this case, we are not just selecting a property but considering all the ItemStack and Item properties.