fix(@clayui/shared): fix focus control collision when having FocusScope nested
There are some cases where FocusScope is declared more than once nested, for example, TimePicker uses FocusScope to manage focus with left and right keys and the component is declared in DatePicker which also has a FocusScope declared.

There is a conflict of who is managing an element aria between the two FocusScopes declared nested and each one can have a configuration, to avoid having conflict between the focus managers, we use the context to know if there is a focus manager parent and we avoid propagation of the event.