chore(@clayui/core): moves the state of keys intermediate to Set<Key>
The reflection of the `layoutKeys` structure is referring to what is rendered in the DOM and built as the items are being rendered, so it is like a stream and on-demand.

To revalidate or rehydrate the items for intermediate we need the `layoutKeys` to have all the data of the structure but that is not the purpose of this structure, so instead we separate the intermediate state to a `Set<Key>` that works independently of the `layoutKeys`.