fix(@clayui/core): fixes the error of not revalidating the intermediate state of the selected item's parents in the first render and expands the items
This commit fixes the error that when starting the rendering of the TreeView with selected items, it was not keeping the state of intermediate, that's because this data is not saved in the backend and just for visual effect, so we need to check which items should have the state of intermediate.

Also with items selected in depth were not visible, we need to expand their parent items to make them visible.

Correcting this implied having to traverse the items tree to know which items should be expanded, we have some conditions that can optimize this so that it does not traverse the entire tree, first, we traverse the tree in search of the selected items and stop traversing until we find all the items, depending on the data and where the selected items are, this can be very fast but this is just an inference that there is no way to predict.

The problem with this is that it can block initial rendering until this operation is finished if it is called at `expandedKeys` initialization so that the TreeView component renders first with all items expanded and with the intermediate state will degrade performance to the first startup, this, when you have a lot of data, can become a problem, with little data it might not be a very noticeable problem.

We also have the option of doing this operation after the TreeView is rendered, so it does not block the first rendering but it will be noticeable for the user to see two changes in the TreeView, first, the TreeView being rendered and then the items being expanded and marked with the state of intermediate.

That's why we're introducing the `selectionHydrationMode` prop here so that developers can choose which behavior they want, for cases where very large data is rendered, the `render-first` option is recommended to avoid blocking the first render. By default, we are leaving `hydrate-first` as it doesn't have the effect of seeing items expand.