fix(@clayui/core): fix error when showing expander when `nestedKey` does not exist in the item
The Expander was only visible when the `children` exists in the item and when it is not empty but it was becoming visible when the `nestedKey` property does not exist in the item when the content is dynamic.

This fix just uses another strategy to show the expander when the content is static.