feat(@clayui/core): adds new properties to set the initial value of properties that can be controlled
We are adding properties to set the initial value of `items`, `selectedKeys`, and `expandedKeys`. According to the feedbacks using the same properties to be initialized and controlled is causing confusion.

With this change most of the cases will be using the initial values property, this will be a breaking change for this component but as it is still in beta we can make changes like this but we want to avoid it but it also means it can be a better DX for the component to avoid confusion.

Also added warnings when using controlled properties incorrectly, the message is the same as what React renders in the console when dealing with controlled and uncontrolled input.