feat(@clayui/modal): improves ClayModal component accessibility
It basically adds attributes that we haven't added for accessibility like `role="dialog"` and others.

We are also moving the focus to the `role="dialog"` which has `aria-labelledby` which can help the screen reader read the dialog title, this avoids escaping the Modal focus control along with `aria- hidden="true"`, this implementation does not break `autoFocus` if used.