chore: updates test threshold for ColorPicker
We recently added new behaviors to ColorPicker, #4691 and #4738, and new tests were created to cover these new scenarios in #4746 and #4781. This increases our current coverage and to continue with our strategy of organic growth and safe testing, we raised the threshold closer to the current one to prevent us from losing coverage or failing to test the new behaviors.