feat(@clayui/form): standardize the controlled and uncontrolled ClayRadioGroup component behavior to `value` prop
We are applying the [RFC 0002](https://github.com/liferay/clay/blob/4bad9532a7ee689b6add7ba8fa1dbda1fe8b37b3/docs/proposals/0002-controlled-and-uncontrolled.md) for the RadioGroup component.

For the Checkbox, Toggle, and Select components, keep the `onChange` behavior with the event argument, we couldn't do a backwards compatibility with the current implementation, so we'll leave this change to v4.