chore(@clayui/shared): refactors test cases using Portal
We are having a problem with the FocusScope reference, apparently, the FocusScope executed by DropDown is not the same as what is being executed in the tests, this makes the contexts that resolve the conflict of focus of the nested manager not work because they are different references. Probably a problem related to the resolution of files in jest that we are configuring badly but for now I am reproducing the test components differently to simulate the case if we had more FocusScope nested.