fix(@clayui/date-picker): fix date picker error not starting with default values using controlled or uncontrolled
Well, this is not a specific bug of the implementation of controlled or uncontrolled, actually, we were not covering this use case, we just consider the value of the `initialMonth` property instead of considering the value of `value` or `defaultValue`.

Considering that this is a behavior that was from the beginning of the component construction, probably people always used the `initialMonth` to start the default value, maybe this was misinterpreted the use of this property that ended up covering a wrong behavior of the component, with the introduction of more solid controlled and uncontrolled components got more clearer for developers.