feat(@clayui/data-provider): adds integration with React.Suspense and ErrorBoundary
This commit adds the initial integration with React.Suspense for data fetching.

This is not a final integration because we need to refactor the cache implementation to be global and move the data from PROMISES and CURSORS to a global instance to simplify the implementation, we are using PROMISES as a data source to retrieve the values in some moments.

Since is integrated with Suspense and there is a promise in progress the component loses any state value or something that is with `useRef`, so we need to move the cache to work global.